+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.4.0-RELEASE
 
 1. Add Sentry breadcrumbs
+2. Add Makefile and Dockerfile for arm machines
 
 ## 1.3.0-RELEASE (2022-01-13)