+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 1.4.0-RELEASE
 
 1. Add Sentry breadcrumbs
-2. Add notion meeting command
+2. Add Makefile and Dockerfile for arm machines
+3. Add notion meeting command
+
 
 ## 1.3.0-RELEASE (2022-01-13)