+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## 1.4.0-RELEASE
 
-1. Add SquadUp feature (/squad up command) - tool to assemble project teams
+1. Add Sentry breadcrumbs
+2. Add Makefile and Dockerfile for arm machines
+3. Add notion meeting command
+4. Add SquadUp feature (/squad up command) - tool to assemble project teams
+
 
 ## 1.3.0-RELEASE (2022-01-13)