+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 2. Add Makefile and Dockerfile for arm machines
 3. Add notion meeting command
 
+## 1.3.1-RELEASE (2022-02-03)
+
+1. Feb coordinape distribution round link update
 
 ## 1.3.0-RELEASE (2022-01-13)